refactor(fossid-webapp): Introduce the concept of an event handler
This will allow, in a future commit, to introduce another mode for the
FossID scanner where the source code to scan is uploaded as a single
archive.
