+++ cli/src/main/resources/logback.xml
@@ -38,4 +38,5 @@
     <logger name="org.ossreviewtoolkit.plugins.scanners.fossid.FossId" level="INFO" />
     <logger name="org.ossreviewtoolkit.plugins.scanners.fossid.FossIdConfig" level="INFO" />
     <logger name="org.ossreviewtoolkit.plugins.scanners.fossid.FossIdUrlProvider" level="INFO" />
+    <logger name="org.ossreviewtoolkit.plugins.scanners.fossid.events.CloneRepositoryHandler" level="INFO" />
 </configuration>
